generator: fix iphdr protocol field size
iphdr.protocol was read as a 32-bits value, while it's actually a 8-bits
value. Fix this issue when reading/writing the L4 protocol.

+++ src/generator/program.h
@@ -115,7 +115,7 @@ struct bf_program_context
 
     /** Layer 4 protocol. Set when the L3 header is processed. Used to define
      * how many bytes to read when processing the packet. */
-    uint16_t l4_proto;
+    uint8_t l4_proto;
 
     /** Layer 2 header. */
     union
